🍻 Copy `amend_user_repository_access` script
